+++ artio-system-tests/src/test/java/uk/co/real_logic/artio/system_tests/FakeOtfAcceptor.java
@@ -96,7 +96,7 @@ public class FakeOtfAcceptor implements OtfMessageAcceptor
 
     public boolean onError(
         final ValidationError error,
-        final int messageType,
+        final long messageType,
         final int tagNumber,
         final AsciiFieldFlyweight value)
     {
